proxy/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
databricks-sdk>=0.30.0
